+++ config/config.exs
@@ -7,5 +7,6 @@ end
 if Mix.env() == :test do
   config :ua_inspector,
     database_path: Path.expand("../test/fixtures", __DIR__),
-    ets_cleanup_delay: 10
+    ets_cleanup_delay: 10,
+    skip_download_readme: true
 end